Change name of 16 col grid system
