+++ app/ids-grid/example.html
@@ -1,137 +1,151 @@
 <ids-grid-container auto="true">
-  <ids-grid-tile>
+  <ids-grid-cell>
     <ids-label font-size="12">12/8/4 Column Grid</ids-label>
-  </ids-grid-tile>
+  </ids-grid-cell>
 </ids-grid-container>
 
 <ids-grid-container>
-  <ids-grid-tile fill="true" span="1">
+  <ids-grid-cell fill="true" span="1">
     <ids-label font-size="12">1</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="1">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="1">
     <ids-label font-size="12">1</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="1">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="1">
     <ids-label font-size="12">1</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="1">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="1">
     <ids-label font-size="12">1</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="1">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="1">
     <ids-label font-size="12">1</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="1">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="1">
     <ids-label font-size="12">1</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="1">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="1">
     <ids-label font-size="12">1</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="1">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="1">
     <ids-label font-size="12">1</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="1">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="1">
     <ids-label font-size="12">1</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="1">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="1">
     <ids-label font-size="12">1</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="1">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="1">
     <ids-label font-size="12">1</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="1">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="1">
     <ids-label font-size="12">1</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="1">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="1">
     <ids-label font-size="12">1</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="1">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="1">
     <ids-label font-size="12">1</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="1">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="1">
     <ids-label font-size="12">1</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="1">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="1">
     <ids-label font-size="12">1</ids-label>
-  </ids-grid-tile>
+  </ids-grid-cell>
 </ids-grid-container>
 
 <ids-grid-container>
-  <ids-grid-tile fill="true" span="2">
+  <ids-grid-cell fill="true" span="2">
     <ids-label font-size="12">2</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="2">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="2">
     <ids-label font-size="12">2</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="2">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="2">
     <ids-label font-size="12">2</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="2">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="2">
     <ids-label font-size="12">2</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="2">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="2">
     <ids-label font-size="12">2</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="2">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="2">
     <ids-label font-size="12">2</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="2">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="2">
     <ids-label font-size="12">2</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="2">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="2">
     <ids-label font-size="12">2</ids-label>
-  </ids-grid-tile>
+  </ids-grid-cell>
 </ids-grid-container>
 
 <ids-grid-container>
-  <ids-grid-tile fill="true" span="4">
+  <ids-grid-cell fill="true" span="4">
     <ids-label font-size="12">4</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="4">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="4">
     <ids-label font-size="12">4</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="4">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="4">
     <ids-label font-size="12">4</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="4">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="4">
     <ids-label font-size="12">4</ids-label>
-  </ids-grid-tile>
+  </ids-grid-cell>
 </ids-grid-container>
 
 <ids-grid-container>
-  <ids-grid-tile fill="true" span="8">
+  <ids-grid-cell fill="true" span="8">
     <ids-label font-size="12">8</ids-label>
-  </ids-grid-tile>
-  <ids-grid-tile fill="true" span="8">
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="8">
     <ids-label font-size="12">8</ids-label>
-  </ids-grid-tile>
+  </ids-grid-cell>
 </ids-grid-container>
 
 <ids-grid-container>
-  <ids-grid-tile fill="true" span="16">
+  <ids-grid-cell fill="true" span="16">
     <ids-label font-size="12">16</ids-label>
-  </ids-grid-tile>
+  </ids-grid-cell>
 </ids-grid-container>
 
 <ids-grid-container auto="true">
-  <ids-grid-tile span="12">
+  <ids-grid-cell>
+    <ids-label font-size="12">Example Layout</ids-label>
+  </ids-grid-cell>
+</ids-grid-container>
+
+<ids-grid-container>
+  <ids-grid-cell fill="true" span="4">
+    <ids-label>Side Content</ids-label>
+    <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Consequuntur similique quis iusto, excepturi odit omnis. Soluta sunt laborum facilis consequatur enim ab et sapiente dignissimos tenetur. Mollitia nam iste aliquam.</p>
+  </ids-grid-cell>
+  <ids-grid-cell fill="true" span="12">
+    <ids-label>Main Content</ids-label>
+    <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Itaque unde eos, tempora eaque fugiat illum ad perspiciatis. Voluptatibus consectetur repellat necessitatibus nihil blanditiis. Sint, impedit! Dolore cupiditate tempora ea doloribus?</p>
+  </ids-grid-cell>
+</ids-grid-container>
+
+<ids-grid-container auto="true">
+  <ids-grid-cell>
     <ids-label font-size="12">Grid Standalone CSS</ids-label>
-  </ids-grid-tile>
+  </ids-grid-cell>
 </ids-grid-container>
 
 <div class="ids-grid">
-  <div class="ids-grid-tile ids-grid-span-4 ids-background-fill">
+  <div class="ids-grid-cell ids-grid-span-4 ids-background-fill">
     <ids-label font-size="12">4 Cols</ids-label>
   </div>
-
-  <div class="ids-grid-tile ids-grid-span-4 ids-background-fill">
+  <div class="ids-grid-cell ids-grid-span-4 ids-background-fill">
     <ids-label font-size="12">4 Cols</ids-label>
   </div>
-
-  <div class="ids-grid-tile ids-grid-span-4 ids-background-fill">
+  <div class="ids-grid-cell ids-grid-span-4 ids-background-fill">
     <ids-label font-size="12">4 Cols</ids-label>
   </div>
-
-  <div class="ids-grid-tile ids-grid-span-4 ids-background-fill">
+  <div class="ids-grid-cell ids-grid-span-4 ids-background-fill">
     <ids-label font-size="12">4 Cols</ids-label>
   </div>
 </div>
