+++ app/ids-grid/index.js
@@ -1,2 +1,2 @@
 import IdsGridContainer from '../../src/ids-grid/ids-grid-container';
-import IdsGridTile from '../../src/ids-grid/ids-grid-tile';
+import IdsGridCell from '../../src/ids-grid/ids-grid-cell';
